Change mail timeout from 300 seconds to 25 seconds
This makes it much easier to detect when there are problems with email configuration. The default timeout in php mailer is 5 minutes, which in some setups can effectively lock the user out for 5 whole minutes due to the session locking if there is a networking issue connecting to the SMTP server (ex: a firewall issue). Additionally, the error will get lost because Apache or load balancers will close the connection after 30 seconds in many cases. By setting the timeout to 25 seconds, we are able to more easily triage errors while still allowing emails plenty of time to go through.